requirements.txt
typer
phidata 